ml-models/requirements.txt

numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
matplotlib==3.7.2
seaborn==0.12.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pyyaml==6.0.1

